--- src/main/java/com/gist/graduation/requirment/domain/GraduationRequirementStatus2.java
@@ -1,30 +1,0 @@
-//package com.gist.graduation.requirment.domain;
-//
-//import com.gist.graduation.requirment.domain.etc.EtcMandatory;
-//import com.gist.graduation.requirment.domain.humanities.Humanities;
-//import com.gist.graduation.requirment.domain.language.LanguageBasic;
-//import com.gist.graduation.requirment.domain.other.OtherUncheckedClass;
-//import com.gist.graduation.requirment.domain.science.ScienceBasic;
-//
-//public enum GraduationRequirementStatus2 {
-//    languageBasic(new LanguageBasic()),
-//    scienceBasic(new ScienceBasic()),
-//    major(new Major()),
-//    humanities(new Humanities()),
-//    etcMandatory(new EtcMandatory()),
-//    OtherUncheckedClass(new OtherUncheckedClass());
-//
-//    private final RequirementStatusBaseEntity statusBaseEntity;
-//
-//    GraduationRequirementStatus2(RequirementStatusBaseEntity statusBaseEntity) {
-//        this.statusBaseEntity = statusBaseEntity;
-//    }
-//
-//    public static checkGraduationRequirements(){
-//        for (GraduationRequirementStatus2 graduationRequirement : values()) {
-//            graduationRequirement.statusBaseEntity.c
-//        }
-//
-//
-//    }
-//}
